extract node lib adapter
